pyproject.toml
[project]
name = "convert-to-md"
version = "0.1.0"
description = "Add your description here"
authors = [{ name = "Donald Thompson", email = "witt3rd@witt3rd.com" }]
dependencies = [
    "fastapi>=0.111.0",
    "aiohttp>=3.9.5",
    "uvicorn>=0.30.0",
    "python-dotenv>=1.0.1",
    "torch>=2.3.0",
    "torchvision>=0.18.0",
    "torchaudio>=2.3.0",
    "marker-pdf>=0.2.13",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.rye.scripts]
srv = "uvicorn convert_to_md.main:app --reload"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/convert_to_md"]
